# perf: cut allocations