@das7pad 2023-09-04
> @das7pad looks like this project is active again, would you be able to rebase this PR?

Sure, done :)

Below are the latest benchmark results

<details>

<summary> mux project benchmarks </summary>

You can reproduce these benchmarks using docker, pinned to CPU 1:
```
$ docker run --rm --pull always -v /logs:/logs --cpuset-cpus 1 -d golang:1.21 bash -exc 'git clone https://github.com/das7pad/mux.git && cd mux && for branch in baseline perf-cut-allocations; do git checkout "$branch" && go test -benchmem -bench . -count 100 -timeout 1h > "/logs/$branch-all.txt"; done; go install golang.org/x/perf/cmd/benchstat@latest; benchstat /logs/baseline-all.txt /logs/perf-cut-allocations-all.txt > /logs/compare-all.txt'
```

<details>
<summary> Modern Xeon E, 3.4 GHz </summary>

``` 
goos: linux
goarch: amd64
pkg: github.com/gorilla/mux
cpu: Intel(R) Xeon(R) E-2278G CPU @ 3.40GHz
                                    │ /logs/baseline-all.txt │  /logs/perf-cut-allocations-all.txt  │
                                    │         sec/op         │   sec/op     vs base                 │
Mux                                             1057.0n ± 0%   943.4n ± 0%  -10.75% (n=100)
MuxSimple/default                                659.1n ± 0%   328.2n ± 0%  -50.20% (n=100)
MuxSimple/omit_route_from_ctx                    653.8n ± 0%   158.3n ± 0%  -75.79% (n=100)
MuxAlternativeInRegexp                           1.588µ ± 0%   1.353µ ± 0%  -14.77% (n=100)
ManyPathVariables                                1.543µ ± 0%   1.457µ ± 0%   -5.61% (n=100)
PopulateContext/no_populated_vars                665.6n ± 0%   329.6n ± 0%  -50.48% (n=100)
PopulateContext/empty_var                        929.0n ± 0%   796.1n ± 0%  -14.31% (n=100)
PopulateContext/populated_vars                   959.3n ± 1%   839.4n ± 1%  -12.51% (n=100)
PopulateContext/omit_route_/static               665.0n ± 0%   159.1n ± 1%  -76.07% (n=100)
PopulateContext/omit_route_/dynamic              923.6n ± 0%   758.3n ± 0%  -17.91% (n=100)
_findQueryKey/0                                  143.8n ± 0%   146.2n ± 0%   +1.67% (p=0.000 n=100)
_findQueryKey/1                                  227.2n ± 0%   227.8n ± 0%        ~ (p=0.058 n=100)
_findQueryKey/2                                  796.7n ± 0%   806.3n ± 1%   +1.20% (p=0.000 n=100)
_findQueryKey/3                                  888.2n ± 0%   893.9n ± 0%   +0.65% (p=0.021 n=100)
_findQueryKey/4                                  3.690n ± 0%   3.722n ± 0%   +0.85% (p=0.000 n=100)
_findQueryKeyGoLib/0                             776.5n ± 0%   752.4n ± 0%   -3.10% (n=100)
_findQueryKeyGoLib/1                             405.5n ± 0%   398.3n ± 0%   -1.76% (n=100)
_findQueryKeyGoLib/2                             2.549µ ± 0%   2.513µ ± 0%   -1.39% (p=0.000 n=100)
_findQueryKeyGoLib/3                             3.355µ ± 0%   3.295µ ± 0%   -1.79% (n=100)
_findQueryKeyGoLib/4                             3.569n ± 0%   3.724n ± 0%   +4.34% (p=0.000 n=100)
geomean                                          474.4n        368.4n       -22.34%

                                    │ /logs/baseline-all.txt │   /logs/perf-cut-allocations-all.txt    │
                                    │          B/op          │     B/op      vs base                   │
Mux                                            1024.0 ± 0%       768.0 ± 0%  -25.00% (n=100)
MuxSimple/default                               720.0 ± 0%       352.0 ± 0%  -51.11% (n=100)
MuxSimple/omit_route_from_ctx                  720.00 ± 0%       48.00 ± 0%  -93.33% (n=100)
MuxAlternativeInRegexp                        2.000Ki ± 0%     1.500Ki ± 0%  -25.00% (n=100)
ManyPathVariables                              1230.0 ± 0%       972.0 ± 0%  -20.98% (n=100)
PopulateContext/no_populated_vars               720.0 ± 0%       352.0 ± 0%  -51.11% (n=100)
PopulateContext/empty_var                      1040.0 ± 0%       784.0 ± 0%  -24.62% (n=100)
PopulateContext/populated_vars                 1024.0 ± 0%       768.0 ± 0%  -25.00% (n=100)
PopulateContext/omit_route_/static             720.00 ± 0%       48.00 ± 0%  -93.33% (n=100)
PopulateContext/omit_route_/dynamic            1040.0 ± 0%       736.0 ± 0%  -29.23% (n=100)
_findQueryKey/0                                 0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/1                                 40.00 ± 0%       40.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/2                                 483.0 ± 0%       483.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/3                                 543.0 ± 0%       543.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/4                                 0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/0                            864.0 ± 0%       864.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/1                            432.0 ± 0%       432.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/2                          1.506Ki ± 0%     1.506Ki ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/3                          1.938Ki ± 0%     1.938Ki ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/4                            0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
geomean                                                    ²                 -34.86%                 ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                    │ /logs/baseline-all.txt │  /logs/perf-cut-allocations-all.txt   │
                                    │       allocs/op        │ allocs/op   vs base                   │
Mux                                             8.000 ± 0%     7.000 ± 0%  -12.50% (n=100)
MuxSimple/default                               7.000 ± 0%     3.000 ± 0%  -57.14% (n=100)
MuxSimple/omit_route_from_ctx                   7.000 ± 0%     1.000 ± 0%  -85.71% (n=100)
MuxAlternativeInRegexp                          16.00 ± 0%     14.00 ± 0%  -12.50% (n=100)
ManyPathVariables                               12.00 ± 0%     11.00 ± 0%   -8.33% (n=100)
PopulateContext/no_populated_vars               7.000 ± 0%     3.000 ± 0%  -57.14% (n=100)
PopulateContext/empty_var                       9.000 ± 0%     8.000 ± 0%  -11.11% (n=100)
PopulateContext/populated_vars                  8.000 ± 0%     7.000 ± 0%  -12.50% (n=100)
PopulateContext/omit_route_/static              7.000 ± 0%     1.000 ± 0%  -85.71% (n=100)
PopulateContext/omit_route_/dynamic             9.000 ± 0%     7.000 ± 0%  -22.22% (n=100)
_findQueryKey/0                                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/1                                 3.000 ± 0%     3.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/2                                 10.00 ± 0%     10.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/3                                 11.00 ± 0%     11.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/4                                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/0                            8.000 ± 0%     8.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/1                            4.000 ± 0%     4.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/2                            24.00 ± 0%     24.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/3                            28.00 ± 0%     28.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/4                            0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=100) ¹
geomean                                                    ²               -27.54%                 ²
¹ all samples are equal
² summaries must be >0 to compute geomean
``` 

---

</details>

<details>
<summary> Older Xeon Gold, 2.3 GHz </summary>

(Actual CPU identifier is rather `Intel(R) Xeon(R) Gold 5122 CPU @ 2.30GHz`)

```
goos: linux
goarch: amd64
pkg: github.com/gorilla/mux
cpu: Intel Xeon Processor (Skylake, IBRS)
                                    │ /logs/baseline-all.txt │  /logs/perf-cut-allocations-all.txt  │
                                    │         sec/op         │   sec/op     vs base                 │
Mux                                              2.615µ ± 2%   2.408µ ± 3%   -7.93% (p=0.000 n=100)
MuxSimple/default                               1600.5n ± 3%   798.9n ± 2%  -50.08% (n=100)
MuxSimple/omit_route_from_ctx                   1549.5n ± 2%   399.8n ± 2%  -74.20% (n=100)
MuxAlternativeInRegexp                           3.703µ ± 2%   3.187µ ± 3%  -13.95% (n=100)
ManyPathVariables                                3.924µ ± 4%   3.591µ ± 3%   -8.50% (p=0.000 n=100)
PopulateContext/no_populated_vars               1572.5n ± 2%   825.7n ± 3%  -47.49% (n=100)
PopulateContext/empty_var                        2.237µ ± 2%   1.871µ ± 3%  -16.36% (n=100)
PopulateContext/populated_vars                   2.352µ ± 3%   1.973µ ± 4%  -16.10% (n=100)
PopulateContext/omit_route_/static              1517.5n ± 3%   394.6n ± 4%  -74.00% (n=100)
PopulateContext/omit_route_/dynamic              2.182µ ± 3%   1.776µ ± 3%  -18.63% (n=100)
_findQueryKey/0                                  337.6n ± 3%   351.6n ± 3%   +4.12% (p=0.005 n=100)
_findQueryKey/1                                  462.6n ± 1%   467.6n ± 2%        ~ (p=0.660 n=100)
_findQueryKey/2                                  1.681µ ± 2%   1.709µ ± 1%        ~ (p=0.093 n=100)
_findQueryKey/3                                  1.833µ ± 1%   1.869µ ± 2%   +1.94% (p=0.011 n=100)
_findQueryKey/4                                  9.438n ± 4%   8.826n ± 2%   -6.48% (p=0.000 n=100)
_findQueryKeyGoLib/0                             1.736µ ± 2%   1.790µ ± 3%   +3.11% (p=0.000 n=100)
_findQueryKeyGoLib/1                             876.6n ± 1%   865.2n ± 2%        ~ (p=0.156 n=100)
_findQueryKeyGoLib/2                             5.473µ ± 1%   5.481µ ± 2%        ~ (p=0.663 n=100)
_findQueryKeyGoLib/3                             7.538µ ± 1%   7.252µ ± 2%   -3.79% (p=0.000 n=100)
_findQueryKeyGoLib/4                             8.777n ± 2%   9.817n ± 4%  +11.86% (p=0.000 n=100)
geomean                                          1.098µ        863.1n       -21.38%

                                    │ /logs/baseline-all.txt │   /logs/perf-cut-allocations-all.txt    │
                                    │          B/op          │     B/op      vs base                   │
Mux                                            1024.0 ± 0%       768.0 ± 0%  -25.00% (n=100)
MuxSimple/default                               720.0 ± 0%       352.0 ± 0%  -51.11% (n=100)
MuxSimple/omit_route_from_ctx                  720.00 ± 0%       48.00 ± 0%  -93.33% (n=100)
MuxAlternativeInRegexp                        2.000Ki ± 0%     1.500Ki ± 0%  -25.00% (n=100)
ManyPathVariables                              1235.0 ± 0%       980.0 ± 0%  -20.65% (n=100)
PopulateContext/no_populated_vars               720.0 ± 0%       352.0 ± 0%  -51.11% (n=100)
PopulateContext/empty_var                      1040.0 ± 0%       784.0 ± 0%  -24.62% (n=100)
PopulateContext/populated_vars                 1024.0 ± 0%       768.0 ± 0%  -25.00% (n=100)
PopulateContext/omit_route_/static             720.00 ± 0%       48.00 ± 0%  -93.33% (n=100)
PopulateContext/omit_route_/dynamic            1040.0 ± 0%       736.0 ± 0%  -29.23% (n=100)
_findQueryKey/0                                 0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/1                                 40.00 ± 0%       40.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/2                                 483.0 ± 0%       483.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/3                                 543.0 ± 0%       543.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/4                                 0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/0                            864.0 ± 0%       864.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/1                            432.0 ± 0%       432.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/2                          1.506Ki ± 0%     1.506Ki ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/3                          1.938Ki ± 0%     1.938Ki ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/4                            0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
geomean                                                    ²                 -34.85%                 ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                    │ /logs/baseline-all.txt │  /logs/perf-cut-allocations-all.txt   │
                                    │       allocs/op        │ allocs/op   vs base                   │
Mux                                             8.000 ± 0%     7.000 ± 0%  -12.50% (n=100)
MuxSimple/default                               7.000 ± 0%     3.000 ± 0%  -57.14% (n=100)
MuxSimple/omit_route_from_ctx                   7.000 ± 0%     1.000 ± 0%  -85.71% (n=100)
MuxAlternativeInRegexp                          16.00 ± 0%     14.00 ± 0%  -12.50% (n=100)
ManyPathVariables                               12.00 ± 0%     11.00 ± 0%   -8.33% (n=100)
PopulateContext/no_populated_vars               7.000 ± 0%     3.000 ± 0%  -57.14% (n=100)
PopulateContext/empty_var                       9.000 ± 0%     8.000 ± 0%  -11.11% (n=100)
PopulateContext/populated_vars                  8.000 ± 0%     7.000 ± 0%  -12.50% (n=100)
PopulateContext/omit_route_/static              7.000 ± 0%     1.000 ± 0%  -85.71% (n=100)
PopulateContext/omit_route_/dynamic             9.000 ± 0%     7.000 ± 0%  -22.22% (n=100)
_findQueryKey/0                                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/1                                 3.000 ± 0%     3.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/2                                 10.00 ± 0%     10.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/3                                 11.00 ± 0%     11.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/4                                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/0                            8.000 ± 0%     8.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/1                            4.000 ± 0%     4.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/2                            24.00 ± 0%     24.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/3                            28.00 ± 0%     28.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/4                            0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=100) ¹
geomean                                                    ²               -27.54%                 ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```

---

</details>

<details>
<summary> Older Xeon E, 2.4 GHz </summary>

``` 
goos: linux
goarch: amd64
pkg: github.com/gorilla/mux
cpu: Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz
                                    │ /logs/baseline-all.txt │  /logs/perf-cut-allocations-all.txt  │
                                    │         sec/op         │   sec/op     vs base                 │
Mux                                              1.698µ ± 1%   1.524µ ± 1%  -10.25% (n=100)
MuxSimple/default                               1031.5n ± 1%   523.9n ± 1%  -49.21% (n=100)
MuxSimple/omit_route_from_ctx                   1023.0n ± 0%   262.9n ± 0%  -74.31% (n=100)
MuxAlternativeInRegexp                           2.463µ ± 0%   2.143µ ± 1%  -13.01% (n=100)
ManyPathVariables                                2.619µ ± 1%   2.401µ ± 1%   -8.30% (n=100)
PopulateContext/no_populated_vars               1025.0n ± 0%   516.4n ± 0%  -49.62% (n=100)
PopulateContext/empty_var                        1.421µ ± 0%   1.257µ ± 1%  -11.58% (n=100)
PopulateContext/populated_vars                   1.481µ ± 1%   1.298µ ± 1%  -12.33% (n=100)
PopulateContext/omit_route_/static              1026.0n ± 0%   255.3n ± 1%  -75.12% (n=100)
PopulateContext/omit_route_/dynamic              1.427µ ± 1%   1.154µ ± 1%  -19.17% (n=100)
_findQueryKey/0                                  234.3n ± 0%   235.2n ± 0%        ~ (p=0.057 n=100)
_findQueryKey/1                                  373.7n ± 0%   372.4n ± 0%        ~ (p=0.317 n=100)
_findQueryKey/2                                  1.310µ ± 1%   1.308µ ± 1%        ~ (p=0.772 n=100)
_findQueryKey/3                                  1.469µ ± 1%   1.482µ ± 1%        ~ (p=0.097 n=100)
_findQueryKey/4                                  6.343n ± 0%   6.304n ± 0%   -0.61% (p=0.000 n=100)
_findQueryKeyGoLib/0                             1.180µ ± 1%   1.181µ ± 1%        ~ (p=0.674 n=100)
_findQueryKeyGoLib/1                             631.4n ± 1%   624.8n ± 1%   -1.05% (p=0.000 n=100)
_findQueryKeyGoLib/2                             4.064µ ± 1%   4.064µ ± 1%        ~ (p=0.717 n=100)
_findQueryKeyGoLib/3                             5.479µ ± 1%   5.494µ ± 0%        ~ (p=0.535 n=100)
_findQueryKeyGoLib/4                             6.323n ± 0%   6.352n ± 0%   +0.46% (p=0.015 n=100)
geomean                                          759.6n        594.1n       -21.79%

                                    │ /logs/baseline-all.txt │   /logs/perf-cut-allocations-all.txt    │
                                    │          B/op          │     B/op      vs base                   │
Mux                                            1024.0 ± 0%       768.0 ± 0%  -25.00% (n=100)
MuxSimple/default                               720.0 ± 0%       352.0 ± 0%  -51.11% (n=100)
MuxSimple/omit_route_from_ctx                  720.00 ± 0%       48.00 ± 0%  -93.33% (n=100)
MuxAlternativeInRegexp                        2.000Ki ± 0%     1.500Ki ± 0%  -25.00% (n=100)
ManyPathVariables                              1254.5 ± 2%       998.0 ± 0%  -20.45% (n=100)
PopulateContext/no_populated_vars               720.0 ± 0%       352.0 ± 0%  -51.11% (n=100)
PopulateContext/empty_var                      1040.0 ± 0%       784.0 ± 0%  -24.62% (n=100)
PopulateContext/populated_vars                 1024.0 ± 0%       768.0 ± 0%  -25.00% (n=100)
PopulateContext/omit_route_/static             720.00 ± 0%       48.00 ± 0%  -93.33% (n=100)
PopulateContext/omit_route_/dynamic            1040.0 ± 0%       736.0 ± 0%  -29.23% (n=100)
_findQueryKey/0                                 0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/1                                 40.00 ± 0%       40.00 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/2                                 483.0 ± 0%       483.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/3                                 543.0 ± 0%       543.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKey/4                                 0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/0                            864.0 ± 0%       864.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/1                            432.0 ± 0%       432.0 ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/2                          1.506Ki ± 0%     1.506Ki ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/3                          1.938Ki ± 0%     1.938Ki ± 0%        ~ (p=1.000 n=100) ¹
_findQueryKeyGoLib/4                            0.000 ± 0%       0.000 ± 0%        ~ (p=1.000 n=100) ¹
geomean                                                    ²                 -34.84%                 ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                    │ /logs/baseline-all.txt │  /logs/perf-cut-allocations-all.txt   │
…